+++ components/collector/src/source_collectors/gitlab.py
@@ -1,23 +1,23 @@
-"""Gitlab metric source."""
+"""GitLab metric source."""
 
 from abc import ABC
 from datetime import datetime
-from typing import cast, List, Optional, Tuple
+from typing import cast, Iterator, List, Optional, Set, Tuple
 from urllib.parse import quote
 
 from dateutil.parser import parse
 import requests
 
 from collector_utilities.functions import days_ago
-from collector_utilities.type import Jobs, Entities, Response, Responses, URL, Value
+from collector_utilities.type import Job, Entities, Response, Responses, URL, Value
 from .source_collector import SourceCollector, UnmergedBranchesSourceCollector
 
 
-class GitlabBase(SourceCollector, ABC):  # pylint: disable=abstract-method
-    """Baseclass for Gitlab collectors."""
+class GitLabBase(SourceCollector, ABC):  # pylint: disable=abstract-method
+    """Base class for GitLab collectors."""
 
     def _gitlab_api_url(self, api: str) -> URL:
-        """Return a Gitlab API url with private token, if present in the parameters."""
+        """Return a GitLab API url with private token, if present in the parameters."""
         url = super()._api_url()
         project = self._parameter("project", quote=True)
         api_url = f"{url}/api/v4/projects/{project}/{api}"
@@ -31,8 +31,8 @@ class GitlabBase(SourceCollector, ABC):  # pylint: disable=abstract-method
         return None  # The private token is passed as URI parameter
 
 
-class GitlabFailedJobs(GitlabBase):
-    """Collector class to get failed job counts from Gitlab."""
+class GitLabJobsBase(GitLabBase):
+    """Base class for GitLab job collectors."""
 
     def _api_url(self) -> URL:
         return self._gitlab_api_url("jobs")
@@ -40,19 +40,48 @@ class GitlabFailedJobs(GitlabBase):
     def _parse_source_responses(self, responses: Responses) -> Tuple[Value, Value, Entities]:
         entities = [
             dict(
-                key=job["id"], name=job["ref"], url=job["web_url"], build_status=job["status"],
-                build_date=str((build_date := parse(job["created_at"])).date()),
+                key=job["id"], name=job["name"], url=job["web_url"], build_status=job["status"], branch=job["ref"],
+                stage=job["stage"], build_date=str((build_date := parse(job["created_at"])).date()),
                 build_age=str(days_ago(build_date)))
-            for job in self.__failed_jobs(responses)]
+            for job in self.__jobs(responses)]
         return str(len(entities)), "100", entities
 
-    @staticmethod
-    def __failed_jobs(responses: Responses) -> Jobs:
-        """Return the failed jobs."""
-        return [job for response in responses for job in response.json() if job["status"] == "failed"]
+    def __jobs(self, responses: Responses) -> Iterator[Job]:
+        """Return the jobs to count."""
+        jobs_seen: Set[Tuple[str, str, str]] = set()
+        for response in responses:
+            for job in response.json():
+                job_fingerprint = job["name"], job["stage"], job["ref"]
+                if job_fingerprint in jobs_seen:
+                    continue
+                jobs_seen.add(job_fingerprint)
+                if self._count_job(job):
+                    yield job
+
+    def _count_job(self, job: Job) -> bool:  # pylint: disable=no-self-use,unused-argument
+        """Return whether to count the job."""
+        raise NotImplementedError  # pragma: nocover
 
 
-class GitlabSourceUpToDateness(GitlabBase):
+class GitLabFailedJobs(GitLabJobsBase):
+    """Collector class to get failed job counts from GitLab."""
+
+    def _count_job(self, job: Job) -> bool:
+        """Return whether the job has failed."""
+        failure_types = list(self._parameter("failure_type"))
+        return job["status"] in failure_types
+
+
+class GitLabUnusedJobs(GitLabJobsBase):
+    """Collector class to get unused job counts from GitLab."""
+
+    def _count_job(self, job: Job) -> bool:
+        """Return whether the job is unused."""
+        max_days = int(cast(str, self._parameter("inactive_job_days")))
+        return days_ago(parse(job["created_at"])) > max_days
+
+
+class GitLabSourceUpToDateness(GitLabBase):
     """Collector class to measure the up-to-dateness of a repo or folder/file in a repo."""
 
     def _api_url(self) -> URL:
@@ -71,7 +100,7 @@ class GitlabSourceUpToDateness(GitlabBase):
             """Get the commits of files recursively."""
             tree_api = self._gitlab_api_url(
                 f"repository/tree?path={file_path}&ref={self._parameter('branch', quote=True)}")
-            tree_response = super(GitlabSourceUpToDateness, self)._get_source_responses(tree_api)[0]
+            tree_response = super(GitLabSourceUpToDateness, self)._get_source_responses(tree_api)[0]
             tree_response.raise_for_status()
             tree = tree_response.json()
             file_paths = [quote(item["path"], safe="") for item in tree if item["type"] == "blob"]
@@ -104,7 +133,7 @@ class GitlabSourceUpToDateness(GitlabBase):
         return value, "100", []
 
 
-class GitlabUnmergedBranches(GitlabBase, UnmergedBranchesSourceCollector):
+class GitLabUnmergedBranches(GitLabBase, UnmergedBranchesSourceCollector):
     """Collector class to measure the number of unmerged branches."""
 
     def _api_url(self) -> URL: